Steam engine is now based on PV=NRT
The steam engine is based on the "inflow" of the steam particles and moves according to that. There are no more janky pipe limitations. One thing to note however is the fact that the pipe goes a little bit too far; this could be solved by altering the image of the pipe itself (preferred) or altering the lengths of the pipe rods (could break everything)